+++ src/Container/CHANGELOG.md
@@ -22,6 +22,18 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Nothing
 -->
 
+## [2.0.0] - 2020-06-21
+### Added
+- `ArgumentResolver` class as default implementation of `ResolverInterface`
+- `ReflectionFactory` class creates instance of any existing class and resolve constructor arguments
+- `ReflectionInvoker` class calls any given callable with resolved arguments
+- `ReflectionContainer` class acts like factory for any existing class
+
+### Removed
+- `Container` class does not manage any existing class no more
+- `Container` class does not implements `ResolverInterface` no more
+- Removed `AbstractContainerDecorator` class
+
 ## [1.0.0] - 2020-06-11
 ### Added
 - First release